WP-C: rebase onto B's generator, Phase-1 protocol, AUC autocast fix
- drop the placeholder degradation, use WP-B's Degradation from integration
- ClassificationMetrics.update softmaxes logits.float() so the train AUC is not
  nan under autocast (reported by WP-D)
- train_config: num_epochs 25, batch_size 256 (Phase-1 protocol)
- variant configs for c-twoview-cos1-inst0.1 and c-twoview-cos3
- tools_cos_table.py: per-epoch cos(z_c,z_d) table from a training log
- acceptance test pinned to origin/integration (local ref can be stale)

+++ src/embedding/training.py
@@ -419,7 +419,9 @@ class ClassificationMetrics:
             self.fn[cls] += (~cls_preds & cls_labels).sum().item()
             self.tn[cls] += (~cls_preds & ~cls_labels).sum().item()
 
-        self.all_probs.append(F.softmax(logits, dim=1).detach().cpu())
+        # .float() first: under autocast logits are fp16 and the softmax/roc_auc_score
+        # round-trip yields nan for the train-line AUC (found by WP-D).
+        self.all_probs.append(F.softmax(logits.float(), dim=1).detach().cpu())
         self.all_labels.append(labels.detach().cpu())
 
     def compute_metrics(self) -> dict:
